requirements.txt
aiohttp==3.9.5
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==25.3.0
certifi==2024.12.14
charset-normalizer==3.4.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
environs==14.2.0
fastapi==0.115.6
fastapi_cors==0.0.6
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
google_search_results==2.4.2
greenlet==3.2.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.30.0
idna==3.10
Jinja2==3.1.6
jiter==0.8.2
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.12
langchain-community==0.3.12
langchain-core==0.3.25
langchain-openai==0.2.12
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.3
langchain-upstage==0.4.0
langgraph==0.2.59
langgraph-checkpoint==2.1.0
langgraph-sdk==0.1.72
langsmith==0.2.11
mangum==0.19.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.5
numpy==1.26.4
openai==1.57.4
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pillow==11.3.0
pinecone==5.3.1
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
propcache==0.3.2
pydantic==2.10.3
pydantic-settings==2.10.1
pydantic_core==2.27.1
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==2.7.0
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.41.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.19.1
torch==2.7.1
tqdm==4.67.1
transformers==4.40.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.5.0
uvicorn==0.34.0
wheel==0.44.0
yarl==1.20.1
